refactor: 콘텐츠 편집 상태에 따른 화면을 컴포넌트로 분리
